ai-course-code/chapter03_prompt/intelligent_customer_service/requirements.txt
# 智能客服系统依赖
openai>=1.0.0
python-dotenv>=1.0.0
